Fix some dependencies issues.
p4-validate no longer requires the yaml package by default, but only if
'--dump_hlir' is used. Few people use that option, so it makes sense to
do it this way. When using the option, an error message will be printed
if the module cannot be imported.

p4-graphs now detects if the dot tool is available before trying to
generate the graphs. If the tool is not detected, it will fail
gracefully.

Also some cleanup in the graphs generation code, to avoid code
duplication.

Increased version to 0.9.42