Allow mutually exclusive control flows to invoke the same table
Under some very stringent conditions.

Mechanics: Invokes the recursive paths of a p4_conditional_node with
empty visited sets, and does the check at the end of the conditional
traversal. The check consists of two parts: first checks that a table
in both paths has no next pointer set. And then it checks that the
tables in each path does not exist in the visited set.