Accept algorithm list in field list calculation
The algorithm attribute in a field list calculation can now be a list of
algorithms:
`algorithm { csum16; identity; }`
The old syntax is still supported can still be used:
`algorithm : csum16`

The consequence for a backend is that
p4_field_list_calculation.algorithm can now either be a Python string or
a list. We introduce a new Python attribute for the
p4_field_list_calculation class, algorithms, which is guaranteed to be a
Python list even when the old syntax is used (in which case the list
will have a single element). We recommend that backends use this new
attribute.

Increased version to 0.9.52