Reject '#' unless it is a cpp line directive
Print an error instead of crashing when finding an unexpected '#'
character when parsing (i.e. if not a preprocessor line directive). Most
of the time the preprocessor will reject the program with an "unknown
directive" error message; however it will let through #pragma directives
which it cannot interpret. This means out P4 parser needs to handled
this case.

Increased version to 0.9.54