Keep supporting 'dummy transitions' in parser
Commit d8caf616afcfc398f9ab03cc4d3de67fa34ab20e tried to detect useless
transitions in p4-hlir and would remove them from the parse graph, in
addition to printing a warning. This breaks existing code which relies
on dummy transition to influence deparser generation. With this change,
we keep the warning but no longer remove the transition from the parse
graph which is handed over to the backend.

Also makes sure that the above warnings are displayed when using
p4-validate, p4-graphs or p4-shell.

Increased version to 0.9.57